requirements.txt
flask
marshmallow
flask_smorest
python_dotenv
sqlalchemy
flask-sqlalchemy
psycopg2
flask_jwt_extended
passlib